Extract getConstraintValueFunction function
